biometrico/requirements.txt
flask>=2.3.0
requests>=2.31.0
pyopenssl>=23.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9